refactored frequency score
