+++ tdp_publicdb/sql.py
@@ -32,6 +32,7 @@ create_sample(views, tissue, gene, tissue_data)
 create_gene_sample_score(views, gene, tissue, tissue_data)
 create_gene_sample_score(views, tissue, gene, tissue_data)
 
+# depletion scores
 create_gene_sample_score(views, gene, cellline, cellline_depletion, 'depletion_')
 create_gene_sample_score(views, cellline, gene, cellline_depletion, 'depletion_')
